refactor: reduce duplicated in canvas picker
